# app/api/faction.py
"""Factions router (catalog) — backed by UnitService.

Uses full paths (no shared prefix) because `/factions` and `/subfactions` are
sibling resources. Catalog writes are admin/seed in principle.
"""

from uuid import UUID

from fastapi import APIRouter, Depends, Response, status
from sqlmodel import SQLModel

from app.api.deps import get_current_admin
from app.api.fields import DiceValue, Name, Stat, WeaponCategory, WriteSchema
from app.api.pagination import Page, PageParams, paginate
from app.api.unit import Ability_Read, Weapon_Read, get_unit_service
from app.core.db.models import FACTION_SUBFACTIONS, FactionName
from app.core.services.service_unit import UnitService

router = APIRouter(tags=["catalog"])


class Subfaction_Read(SQLModel):
    id: UUID
    name: str


class Subfaction_ListRead(SQLModel):
    # Flat listing (GET /subfactions) carries faction_id, which the nested form
    # inside Faction_Read can omit because its parent already supplies it.
    id: UUID
    faction_id: UUID
    name: str


class Faction_Read(SQLModel):
    id: UUID
    name: str
    subfactions: list[Subfaction_Read] = []


class Faction_Create(WriteSchema):
    # Restricted to the canonical faction list: an unknown value (e.g. a
    # misspelling) is rejected with 422, and the allowed set is published in the
    # OpenAPI schema (so a frontend can render a dropdown).
    name: FactionName


class Subfaction_Create(WriteSchema):
    faction_id: UUID
    name: Name


class Weapon_Create(WriteSchema):
    name: Name
    category: WeaponCategory
    attacks: DiceValue
    weapon_skill: Stat
    strength: Stat
    armor_piercing: Stat
    damage: DiceValue
    range_inches: Stat | None = None
    keywords: list[str] | None = None


class Ability_Create(WriteSchema):
    name: Name
    description: str


class Weapon_Update(WriteSchema):
    name: Name | None = None
    category: WeaponCategory | None = None
    attacks: DiceValue | None = None
    weapon_skill: Stat | None = None
    strength: Stat | None = None
    armor_piercing: Stat | None = None
    damage: DiceValue | None = None
    range_inches: Stat | None = None
    keywords: list[str] | None = None


class Ability_Update(WriteSchema):
    name: Name | None = None
    description: str | None = None


@router.get("/factions", response_model=Page[Faction_Read])
def list_factions(
    page: PageParams = Depends(),
    service: UnitService = Depends(get_unit_service),
) -> Page[Faction_Read]:
    items = service.list_factions(limit=page.limit, offset=page.offset)
    return paginate(items, service.count_factions(), page)


@router.get("/subfactions", response_model=Page[Subfaction_ListRead])
def list_subfactions(
    faction_id: UUID | None = None,
    page: PageParams = Depends(),
    service: UnitService = Depends(get_unit_service),
) -> Page[Subfaction_ListRead]:
    """Every created subfaction, optionally filtered to one faction.

    Distinct from GET /taxonomy: that publishes the *allowed* subfaction names
    (from the constant); this lists the rows actually created in the DB (with ids).
    """
    items = service.list_subfactions(faction_id, limit=page.limit, offset=page.offset)
    return paginate(items, service.count_subfactions(faction_id), page)


@router.get("/taxonomy", response_model=dict[str, list[str]])
def faction_taxonomy() -> dict[str, list[str]]:
    """The allowed subfactions under each faction (for building dropdowns).

    Sourced from the `FACTION_SUBFACTIONS` map, not the DB — these are the values
    `create_subfaction` will accept, whether or not any have been created yet.

    Deliberately *not* under `/factions/...`: it's static reference data, not a
    faction resource, and `/factions/taxonomy` would collide with a future
    `GET /factions/{faction_id}` (both claim the same path slot). See ROADMAP R12.
    """
    return {faction.value: list(subs) for faction, subs in FACTION_SUBFACTIONS.items()}


@router.post(
    "/factions",
    response_model=Faction_Read,
    status_code=status.HTTP_201_CREATED,
    dependencies=[Depends(get_current_admin)],
)
def create_faction(payload: Faction_Create, service: UnitService = Depends(get_unit_service)) -> Faction_Read:
    return service.create_faction(payload.name)


@router.post(
    "/subfactions",
    response_model=Subfaction_Read,
    status_code=status.HTTP_201_CREATED,
    dependencies=[Depends(get_current_admin)],
)
def create_subfaction(
    payload: Subfaction_Create,
    service: UnitService = Depends(get_unit_service),
) -> Subfaction_Read:
    return service.create_subfaction(payload.faction_id, payload.name)


@router.delete(
    "/subfactions/{subfaction_id}",
    status_code=status.HTTP_204_NO_CONTENT,
    dependencies=[Depends(get_current_admin)],
)
def delete_subfaction(subfaction_id: UUID, service: UnitService = Depends(get_unit_service)) -> Response:
    service.delete_subfaction(subfaction_id)
    return Response(status_code=status.HTTP_204_NO_CONTENT)


@router.post(
    "/weapons",
    response_model=Weapon_Read,
    status_code=status.HTTP_201_CREATED,
    dependencies=[Depends(get_current_admin)],
)
def create_weapon(payload: Weapon_Create, service: UnitService = Depends(get_unit_service)) -> Weapon_Read:
    return service.create_weapon(**payload.model_dump())


@router.get("/weapons", response_model=Page[Weapon_Read])
def list_weapons(
    page: PageParams = Depends(),
    service: UnitService = Depends(get_unit_service),
) -> Page[Weapon_Read]:
    items = service.list_weapons(limit=page.limit, offset=page.offset)
    return paginate(items, service.count_weapons(), page)


@router.patch(
    "/weapons/{weapon_id}",
    response_model=Weapon_Read,
    dependencies=[Depends(get_current_admin)],
)
def update_weapon(
    weapon_id: UUID,
    payload: Weapon_Update,
    service: UnitService = Depends(get_unit_service),
) -> Weapon_Read:
    return service.update_weapon(weapon_id, **payload.model_dump(exclude_unset=True))


@router.delete(
    "/weapons/{weapon_id}",
    status_code=status.HTTP_204_NO_CONTENT,
    dependencies=[Depends(get_current_admin)],
)
def delete_weapon(weapon_id: UUID, service: UnitService = Depends(get_unit_service)) -> Response:
    service.delete_weapon(weapon_id)
    return Response(status_code=status.HTTP_204_NO_CONTENT)


@router.post(
    "/abilities",
    response_model=Ability_Read,
    status_code=status.HTTP_201_CREATED,
    dependencies=[Depends(get_current_admin)],
)
def create_ability(payload: Ability_Create, service: UnitService = Depends(get_unit_service)) -> Ability_Read:
    return service.create_ability(payload.name, payload.description)


@router.get("/abilities", response_model=Page[Ability_Read])
def list_abilities(
    page: PageParams = Depends(),
    service: UnitService = Depends(get_unit_service),
) -> Page[Ability_Read]:
    items = service.list_abilities(limit=page.limit, offset=page.offset)
    return paginate(items, service.count_abilities(), page)


@router.patch(
    "/abilities/{ability_id}",
    response_model=Ability_Read,
    dependencies=[Depends(get_current_admin)],
)
def update_ability(
    ability_id: UUID,
    payload: Ability_Update,
    service: UnitService = Depends(get_unit_service),
) -> Ability_Read:
    return service.update_ability(ability_id, **payload.model_dump(exclude_unset=True))


@router.delete(
    "/abilities/{ability_id}",
    status_code=status.HTTP_204_NO_CONTENT,
    dependencies=[Depends(get_current_admin)],
)
def delete_ability(ability_id: UUID, service: UnitService = Depends(get_unit_service)) -> Response:
    service.delete_ability(ability_id)
    return Response(status_code=status.HTTP_204_NO_CONTENT)